refactor: reorganize AI service classes and update package structure
