+++ backend/src/main/java/ch/nova_omnia/lernello/service/aiClient/ChatRequest.java
@@ -1,4 +1,4 @@
-package ch.nova_omnia.lernello.service.ai;
+package ch.nova_omnia.lernello.service.aiClient;
 
 import java.util.List;
 import java.util.Map;
